refs #9 - Changing LCA function to semantic file
LCA function computes the Lowest Common Ancestor. Since it is used by Type Checker, Type Inferencer and Inferencer Manager, all classes in different files, the location of the function was changed.
Also, note that when using self_type it now changes it for the fixed type of the specific instance of the class